added backendapi func and connected with frontend

+++ InventoryManagement/backend/src/main/java/com/Project/InventoryManagement/RestController/AdminRestController.java
@@ -1,17 +1,20 @@
 package com.Project.InventoryManagement.RestController;
 
 
+import java.util.List;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.Project.InventoryManagement.DTO.ProductDTO;
+import com.Project.InventoryManagement.DTO.SuccessMsg;
 import com.Project.InventoryManagement.Entity.Product;
 import com.Project.InventoryManagement.Entity.Stock;
 import com.Project.InventoryManagement.Entity.Supplier;
@@ -31,43 +34,56 @@ public class AdminRestController extends EmployeeRestController{
         return ResponseEntity.ok(stock); 
     }
 
-    @DeleteMapping("/delete/product/id")
-    public void deleteById(@PathVariable int id){
-        inventoryService.deleteById(id);
-    }
-
     @PostMapping("/save/supplier")
     public ResponseEntity<Supplier> saveSupplier(@RequestBody Supplier supplier){
         supplier = inventoryService.saveSupplier(supplier);
         return ResponseEntity.ok(supplier);
     }
+    @DeleteMapping("/product/delete")
+    public ResponseEntity<SuccessMsg> deleteBySupplierId(@RequestBody Map<String, String> map){
+        String name = map.get("name");
+        String type = map.get("type");
+        int id;
+        try {id = Integer.parseInt(map.get("id"));}
+        catch (Exception e) { id = -1;}
 
-    @PostMapping("/post/supplier/name")
-    public ResponseEntity<Supplier> updateSupplierName(@PathVariable int id, @PathVariable String name){
-        Supplier supplier = inventoryService.updateSupplierName(id, name);
-        return (supplier != null) ? ResponseEntity.ok(supplier) : ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        List<ProductDTO> productDTO;
+        if (id != -1) {
+            inventoryService.deleteByProductId(id);
+            inventoryService.deleteById(id);
+        } else if (name != "" && type != "") {
+            productDTO = inventoryService.getProductsStockByNameAndType(name, type);
+            for (ProductDTO p: productDTO){
+                inventoryService.deleteByProductId(p.getId());
+                inventoryService.deleteById(p.getId());
+            }
+        } else if (name != "") {
+            productDTO = inventoryService.getProductsStockByName(name);
+            for (ProductDTO p: productDTO){
+                inventoryService.deleteByProductId(p.getId());
+                inventoryService.deleteById(p.getId());
+            }
+        } else if (type != "") {
+            productDTO = inventoryService.getProductsStockByType(type);
+            for (ProductDTO p: productDTO){
+                inventoryService.deleteByProductId(p.getId());
+                inventoryService.deleteById(p.getId());
+            }
+        }
+        
+        return ResponseEntity.ok(new SuccessMsg());
     }
 
-    @PostMapping("/post/supplier/no")
-    public ResponseEntity<Supplier> updateSupplierNo(@PathVariable int id, @PathVariable String no){
-        Supplier supplier = inventoryService.updateSupplierNo(id, no);
-        return (supplier != null) ? ResponseEntity.ok(supplier) : ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-    }
+    @DeleteMapping("/supplier/delete")
+    public ResponseEntity<SuccessMsg> deleteSupplier(@RequestBody Map<String, String> map){
+        String name = map.get("name");
+        int id;
+        try {id = Integer.parseInt(map.get("id"));}
+        catch (Exception e) { id = -1;}
+        if (!name.equals("")) inventoryService.deleteBySupplierName(name);
+        else if (id != -1) inventoryService.deleteBySupplierId(id);
 
-    @PostMapping("/post/supplier/email")
-    public ResponseEntity<Supplier> updateSupplierEmail(@PathVariable int id, @PathVariable String email){
-        Supplier supplier = inventoryService.updateSupplierEmail(id, email);
-        return (supplier != null) ? ResponseEntity.ok(supplier) : ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-    }
-
-    @DeleteMapping("/delete/product/supplierId")
-    public void deleteBySupplierId(@PathVariable int id){
-        inventoryService.deleteBySupplierId(id);
-    }
-
-    @DeleteMapping("/delete/supplier/name")
-    public void deleteBySupplierName(@PathVariable String name){
-        inventoryService.deleteBySupplierName(name);
+        return ResponseEntity.ok(new SuccessMsg());
     }
 
     @PostMapping("save/product")
@@ -76,9 +92,63 @@ public class AdminRestController extends EmployeeRestController{
         return ResponseEntity.ok(product);
     }
 
-    @PutMapping("/put/product/name")
-    public ResponseEntity<Product> updateProductName(@PathVariable int id, @PathVariable String name){
-        Product product = inventoryService.updateProductName(id, name);
-        return (product != null) ? ResponseEntity.ok(product) : ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
+    @PostMapping("supplier/update")
+    public ResponseEntity<List<Supplier>> updateSupplier(@RequestBody Map<String, String> map){
+        String name = map.get("name");
+        String field = map.get("field");
+        String value = map.get("value");
+
+        int id;
+        try { 
+            id = Integer.parseInt(map.get("id"));
+        } catch (Exception e) {
+            id = -1;
+        }
+        
+        List<Supplier> suppliers = inventoryService.getSuppliers(id, name);
+        if (suppliers == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        for (Supplier s: suppliers) {
+            if (field.equals("Name")) {
+                inventoryService.updateSupplierName(s.getSupplierId(), value);
+            } else if (field.equals("Phone No")) {
+                inventoryService.updateSupplierNo(s.getSupplierId(), value);
+            } else if (field.equals("Email")) {
+                inventoryService.updateSupplierEmail(s.getSupplierId(), value);
+            } else return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(suppliers);    
+    }
+
+    @PostMapping("/product/update")
+    public ResponseEntity<SuccessMsg> updateProduct(@RequestBody Map<String, String> map){
+        String name = map.get("name");
+        String type = map.get("type");
+        String field = map.get("field");//['Name', 'Type', 'Price', 'Quantity', 'Desc']
+        String value = map.get("value");
+        int id;
+        try { 
+            id = Integer.parseInt(map.get("id"));  
+        } catch (Exception e) { id = -1; }
+
+        List<ProductDTO> productDTO;
+        if (id != -1) {
+            productDTO = inventoryService.getProductStockById(id);
+        } else if (name != "" && type != "") {
+            productDTO = inventoryService.getProductsStockByNameAndType(name, type);
+        } else if (name != "") {
+            productDTO = inventoryService.getProductsStockByName(name);
+        } else if (type != "") {
+            productDTO = inventoryService.getProductsStockByType(type);
+        } else return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        for (ProductDTO p: productDTO) {
+            System.out.println(field);
+            if (field.equals("Name")) inventoryService.updateProductName(p.getId(), value);
+            else if (field.equals("Type")) inventoryService.updateProductType(p.getId(), value);
+            else if (field.equals("Price")) inventoryService.updatePrice(p.getId(), Float.parseFloat(value));
+            else if (field.equals("Quantity")) inventoryService.updateQuantity(p.getId(), Integer.parseInt(value));
+            else if (field.equals("Desc")) inventoryService.updateProductDesc(p.getId(), value);
+        }
+
+        return ResponseEntity.ok(new SuccessMsg());
     }
 }
